solucionado cloudinary, correcciones en swagger

+++ app/Http/Controllers/OrderController.php
@@ -79,6 +79,18 @@ class OrderController extends Controller
     public function destroy(string $id)
     {
         $order = Order::find($id);
+    
+        if (!$order) {
+            return response()->json(['La orden no existe'], 404);
+        }
+
+        $pet = Pet::where('id_order', $order->id)->first();
+        
+        if ($pet) {
+            $pet->id_order = null;
+            $pet->save();
+        }
+        
         $order->delete();
 
         return redirect('/orders');
